% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateMonadCodeGenerator.R
\name{GenerateSetterCode}
\alias{GenerateSetterCode}
\title{Generate State monad setter code.}
\usage{
GenerateSetterCode(monadName, memberName, monadObjectArgumentName = NULL,
  memberClassName = NULL)
}
\arguments{
\item{monadName}{The monad name (to be used as function prefix); a string.}

\item{memberName}{The name of the monad member.}

\item{monadObjectArgumentName}{The monad object name used in the generated functions code.}

\item{memberClassName}{Member class name (lile "character", "numeric", "list", etc.)}
}
\description{
Generate State monad setter code.
}
