% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateMonadCodeGenerator.R
\name{GenerateStateMonadCode}
\alias{GenerateStateMonadCode}
\title{Generate State monad code.}
\usage{
GenerateStateMonadCode(monadName, memberNames = NULL,
  monadObjectArgumentName = NULL, outputFile = NULL)
}
\arguments{
\item{monadName}{The monad name (to be used as function prefix); a string.
If NULL the stencil code name is used.}

\item{memberNames}{A character vector wit the names of the monad members.}

\item{monadObjectArgumentName}{The monad object name used in the generated functions code.}

\item{outputFile}{The file name to contain the generated monad code.}
}
\description{
Generate State monad code.
}
\details{
There are several stencil code files that are read and the corresponding names
replaced. If \code{memberNames} is not NULL then those names are used the setters and takers
are generated and state monad unit S3 object has \code{memberNames} as element names.
If \code{outputFile} (of course) no code is written.
}
