% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateMonadCodeGenerator.R
\name{GenerateTakerCode}
\alias{GenerateTakerCode}
\title{Generate State monad taker code.}
\usage{
GenerateTakerCode(monadName, memberName, monadObjectArgumentName = NULL)
}
\arguments{
\item{monadName}{The monad name (to be used as function prefix); a string.}

\item{memberName}{The name of the monad member.}

\item{monadObjectArgumentName}{The monad object name used in the generated functions code.}
}
\description{
Generate State monad taker code.
}
