% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{textMStateCoreFunctions}
\alias{textMStateCoreFunctions}
\title{State monad core functions stencil.}
\format{Text lines}
\usage{
textMStateCoreFunctions
}
\description{
A list of text lines that define State monad core functions.
}
\keyword{datasets}
