% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{textMStateSetMember}
\alias{textMStateSetMember}
\title{State monad setter function stencil.}
\format{Text lines}
\usage{
textMStateSetMember
}
\description{
A list of text lines that define State monad setter function.
}
\keyword{datasets}
