% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{textMStateTakeMember}
\alias{textMStateTakeMember}
\title{State monad taker function stencil.}
\format{Text lines}
\usage{
textMStateTakeMember
}
\description{
A list of text lines that define State monad setter function.
}
\keyword{datasets}
